split info factory to create lifecycle management class
