+++ dev/io.openliberty.microprofile.telemetry.internal.common/bnd.bnd
@@ -75,7 +75,8 @@ Import-Package: \
 	com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
 	io.openliberty.io.opentelemetry;version=latest,\
 	io.openliberty.org.jboss.resteasy.common;version=latest,\
-	io.opentelemetry.semconv:opentelemetry-semconv;version='${alphaSemconvOpenTelemetryVersion}';strategy=exact
+	io.opentelemetry.semconv:opentelemetry-semconv;version='${alphaSemconvOpenTelemetryVersion}';strategy=exact,\
+	com.ibm.ws.kernel.boot.common
 	
 
 WS-TraceGroup: TELEMETRY
