+++ dev/io.openliberty.microprofile.telemetry.internal.common/src/io/openliberty/microprofile/telemetry/internal/common/info/OpenTelemetryInfoFactoryImpl.java
@@ -17,10 +17,8 @@ import java.security.PrivilegedAction;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiFunction;
 
-import org.apache.commons.lang3.concurrent.LazyInitializer;
 import org.eclipse.microprofile.config.Config;
 import org.eclipse.microprofile.config.ConfigProvider;
 import org.eclipse.microprofile.config.spi.ConfigSource;
@@ -30,14 +28,7 @@ import org.osgi.service.component.annotations.Reference;
 
 import com.ibm.websphere.ras.Tr;
 import com.ibm.websphere.ras.TraceComponent;
-import com.ibm.ws.container.service.app.deploy.ApplicationInfo;
-import com.ibm.ws.container.service.app.deploy.extended.ExtendedApplicationInfo;
-import com.ibm.ws.container.service.metadata.MetaDataSlotService;
-import com.ibm.ws.container.service.state.ApplicationStateListener;
-import com.ibm.ws.container.service.state.StateChangeException;
-import com.ibm.ws.runtime.metadata.ApplicationMetaData;
 import com.ibm.ws.runtime.metadata.ComponentMetaData;
-import com.ibm.ws.runtime.metadata.MetaDataSlot;
 import com.ibm.ws.threadContext.ComponentMetaDataAccessorImpl;
 
 import io.openliberty.microprofile.telemetry.internal.common.AgentDetection;
@@ -50,8 +41,8 @@ import io.opentelemetry.sdk.resources.Resource;
 import io.opentelemetry.sdk.resources.ResourceBuilder;
 
 // We want this to start before CDI so the meta data slot is ready before anyone triggers the CDI producer.
-@Component(service = { ApplicationStateListener.class, OpenTelemetryInfoFactory.class }, property = { "service.vendor=IBM", "service.ranking:Integer=1500" })
-public class OpenTelemetryInfoFactoryImpl implements ApplicationStateListener, OpenTelemetryInfoFactory {
+@Component(service = { OpenTelemetryInfoFactory.class }, property = { "service.vendor=IBM", "service.ranking:Integer=1500" })
+public abstract class OpenTelemetryInfoFactoryImpl implements OpenTelemetryInfoFactory {
 
     /*
      * Class summary.
@@ -73,87 +64,18 @@ public class OpenTelemetryInfoFactoryImpl implements ApplicationStateListener, O
 
     private static final TraceComponent tc = Tr.register(OpenTelemetryInfoFactoryImpl.class);
 
-    private final MetaDataSlot slotForOpenTelemetryInfoHolder;
-    private EnabledOpenTelemetryInfo otelRuntimeInstance = null;
-
     //This contains API calls that change between the upstream open telemetry version.
     //We get a partially configued SDK Builder from OSGi becase we are in a static context
     //and do not know which version of mpTelemetry will be in use.
     private final OpenTelemetryVersionedConfiguration openTelemetryVersionedConfiguration;
 
     @Activate
-    public OpenTelemetryInfoFactoryImpl(@Reference MetaDataSlotService slotService,
-                                        @Reference OpenTelemetryVersionedConfiguration openTelemetryVersionedConfiguration) {
-
-        slotForOpenTelemetryInfoHolder = slotService.reserveMetaDataSlot(ApplicationMetaData.class);
+    public OpenTelemetryInfoFactoryImpl(@Reference OpenTelemetryVersionedConfiguration openTelemetryVersionedConfiguration) {
 
         this.openTelemetryVersionedConfiguration = openTelemetryVersionedConfiguration;
-
-        OpenTelemetry runtimeInstance = this.openTelemetryVersionedConfiguration.createServerOpenTelemetryInfo(getServerTelemetryProperties());
-        if (runtimeInstance != null) {
-            EnabledOpenTelemetryInfo runtimeOpenTelemetryInfo = new EnabledOpenTelemetryInfo(true, runtimeInstance, OpenTelemetryConstants.OTEL_RUNTIME_INSTANCE_NAME);
-            this.otelRuntimeInstance = runtimeOpenTelemetryInfo;
-        }
     }
 
     @Override
-    public OpenTelemetryInfo getOpenTelemetryInfo() {
-
-        try {
-            ApplicationMetaData metaData = null;
-            ComponentMetaData cmd = ComponentMetaDataAccessorImpl.getComponentMetaDataAccessor().getComponentMetaData();
-            if (cmd != null) {
-                metaData = cmd.getModuleMetaData().getApplicationMetaData();
-            }
-            return getOpenTelemetryInfo(metaData);
-        } catch (Exception e) {
-            Tr.error(tc, Tr.formatMessage(tc, "CWMOT5002.telemetry.error", e));
-            return new ErrorOpenTelemetryInfo();
-        }
-    }
-
-    @Override
-    public boolean isRuntimeEnabled() {
-        return otelRuntimeInstance != null;
-    }
-
-    //A shortcut method to avoid fetching metadata more than we need to.
-    @Override
-    public OpenTelemetryInfo getOpenTelemetryInfo(ApplicationMetaData metaData) {
-
-        //Return runtime instance if it exists, otherwise return the app instance.
-        if (this.otelRuntimeInstance != null) {
-            if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-                Tr.debug(tc, "Returning {0} OTEL instance.", OpenTelemetryConstants.OTEL_RUNTIME_INSTANCE_NAME);
-            }
-            return otelRuntimeInstance;
-        }
-
-        if (metaData == null) {
-            return new DisabledOpenTelemetryInfo();
-        }
-
-        try {
-            OpenTelemetryInfoReference atomicRef = (OpenTelemetryInfoReference) metaData.getMetaData(slotForOpenTelemetryInfoHolder);
-            if (atomicRef == null) {
-                //If this is triggered by internal code that isn't supposed to call ApplicationStateListener.applicationStarting() don't throw an error
-                String j2EEName = metaData.getJ2EEName().toString();
-                if (j2EEName.startsWith("io.openliberty") || j2EEName.startsWith("com.ibm.ws")
-                    || j2EEName.startsWith("arquillian-liberty-support")) {
-                    Tr.info(tc, "CWMOT5100.tracing.is.disabled", j2EEName);
-                    return new DisabledOpenTelemetryInfo();
-                }
-                //If it isn't throw something nicer than an NPE.
-                throw new IllegalStateException("Attempted to create openTelemetaryInfo for application " + j2EEName + " which has not gone through ApplicationStarting");
-            }
-            LazyInitializer<OpenTelemetryInfo> supplier = atomicRef.get();
-            return supplier.get();
-        } catch (Exception e) {
-            Tr.error(tc, Tr.formatMessage(tc, "CWMOT5002.telemetry.error", e));
-            return new ErrorOpenTelemetryInfo();
-        }
-    }
-
     public OpenTelemetryInfo createOpenTelemetryInfo() {
         try {
 
@@ -192,11 +114,22 @@ public class OpenTelemetryInfoFactoryImpl implements ApplicationStateListener, O
 
     }
 
-    private static OpenTelemetryInfo createDisposedOpenTelemetryInfo() {
+    @Override
+    public OpenTelemetryInfo createDisposedOpenTelemetryInfo() {
         try {
             String appName = ComponentMetaDataAccessorImpl.getComponentMetaDataAccessor().getComponentMetaData().getJ2EEName().getApplication();
             return new DisposedOpenTelemetryInfo(appName); //Getting appName like this was easier than writing a functional interface
 
+        } catch (Exception e) {
+            Tr.error(tc, Tr.formatMessage(tc, "CWMOT5002.telemetry.error", e));
+            return new ErrorOpenTelemetryInfo();
+        }
+    }
+
+    @Override
+    public OpenTelemetryInfo createDisabledOpenTelemetryInfo() {
+        try {
+            return new DisabledOpenTelemetryInfo();
         } catch (Exception e) {
             Tr.error(tc, Tr.formatMessage(tc, "CWMOT5002.telemetry.error", e));
             return new ErrorOpenTelemetryInfo();
@@ -295,59 +228,6 @@ public class OpenTelemetryInfoFactoryImpl implements ApplicationStateListener, O
         }
     }
 
-    /** {@inheritDoc} */
-    @Override
-    public void applicationStarting(ApplicationInfo appInfo) throws StateChangeException {
-        //We do not actually initilize on application starting, we do that lazily if this is needed.
-
-        //We don't use app scoped OpenTelemetry objects if the server scoped object exists
-        if (otelRuntimeInstance != null) {
-            return;
-        }
-
-        ExtendedApplicationInfo extAppInfo = (ExtendedApplicationInfo) appInfo;
-        OpenTelemetryInfoReference oTelRef = (OpenTelemetryInfoReference) extAppInfo.getMetaData().getMetaData(slotForOpenTelemetryInfoHolder);
-
-        LazyInitializer<OpenTelemetryInfo> newSupplier = LazyInitializer.<OpenTelemetryInfo> builder().setInitializer(this::createOpenTelemetryInfo)
-                                                                        .setCloser(info -> info.dispose()).get();
-
-        if (oTelRef == null) {
-            oTelRef = new OpenTelemetryInfoReference();
-        }
-
-        oTelRef.set(newSupplier);
-        extAppInfo.getMetaData().setMetaData(slotForOpenTelemetryInfoHolder, oTelRef);
-    }
-
-    @Override
-    public void applicationStarted(ApplicationInfo appInfo) throws StateChangeException {} //no-op
-
-    @Override
-    public void applicationStopping(ApplicationInfo appInfo) {} //no-op
-
-    @Override
-    public void applicationStopped(ApplicationInfo appInfo) {
-        //We don't use app scoped OpenTelemetry objects if the server scoped object exists
-        if (otelRuntimeInstance != null) {
-            return;
-        }
-
-        ExtendedApplicationInfo extAppInfo = (ExtendedApplicationInfo) appInfo;
-        OpenTelemetryInfoReference oTelRef = (OpenTelemetryInfoReference) extAppInfo.getMetaData().getMetaData(slotForOpenTelemetryInfoHolder);
-
-        LazyInitializer<OpenTelemetryInfo> newSupplier = LazyInitializer.<OpenTelemetryInfo> builder().setInitializer(OpenTelemetryInfoFactoryImpl::createDisposedOpenTelemetryInfo)
-                                                                        .setCloser(info -> info.dispose()).get();
-
-        LazyInitializer<OpenTelemetryInfo> oldSupplier = oTelRef.getAndSet(newSupplier);
-
-        try {
-            oldSupplier.close();
-        } catch (Exception e) {
-            Tr.warning(tc, "applicationStopped", "failed to dispose of OpenTelemetry");//TODO better message
-        }
-
-    }
-
     //Adds the service name to the resource attributes
     private Resource customizeResource(Resource resource, ConfigProperties c) {
         ResourceBuilder builder = resource.toBuilder();
@@ -376,6 +256,9 @@ public class OpenTelemetryInfoFactoryImpl implements ApplicationStateListener, O
         return serviceName;
     }
 
+    @Override
+    public abstract boolean isRuntimeEnabled();
+
     //Interfaces and private classes only relevent to this factory.
 
     public interface OpenTelemetryVersionedConfiguration {
@@ -383,19 +266,5 @@ public class OpenTelemetryInfoFactoryImpl implements ApplicationStateListener, O
                                          ClassLoader classLoader);
 
         OpenTelemetry createServerOpenTelemetryInfo(HashMap<String, String> hashMap);
-    }
-
-    /*
-     * There are two race conditions we need to protect against.
-     *
-     * On app startup/shutdown we need to swap the holder associated with the application.
-     * We protect against race conditions here by using an AtomicReference.
-     *
-     * Within the context of an application's lifecycle we need to ensure OpenTelemetryInfo
-     * is only created once. LazySupplier handles this.
-     */
-    private class OpenTelemetryInfoReference extends AtomicReference<LazyInitializer<OpenTelemetryInfo>> {
-
-        private static final long serialVersionUID = -4884222080590544495L;
     }
 }
